# backend/openapi_milestone4.yaml
openapi: 3.0.3
info:
  title: PulseTrack HRMS API - Milestone 4
  version: 1.0.0
  description: |
    # PulseTrack HRMS - GenAI-Powered Human Resource Management System
    
    ## Project Information
    - **Team**: Team 11, IIT Madras
    - **Course**: Software Engineering (SEP 2025)
    - **Milestone**: 4 - API Endpoints Documentation
    
    ## Overview
    PulseTrack HRMS is a comprehensive, AI-powered Human Resource Management System designed to revolutionize 
    workforce management by making HR processes smarter, faster, and more human-centric.
    
    ## GenAI Integrations
    This system integrates multiple GenAI services to automate and enhance HR processes:
    
    ### 1. Google Gemini API
    - **Purpose**: Natural language processing, content generation, and intelligent analysis
    - **Used In**: Policy RAG, Resume Screening, Job Description Generation
    - **Model**: gemini-1.5-flash
    - **Features**: Text generation, embedding, multi-turn conversations
    
    ### 2. ChromaDB
    - **Purpose**: Vector database for semantic search
    - **Used In**: Policy RAG for document similarity search
    - **Features**: Persistent storage, similarity search, metadata filtering
    
    ### 3. Langchain
    - **Purpose**: RAG pipeline orchestration
    - **Used In**: Policy RAG question-answering system
    - **Features**: Document loading, text splitting, retrieval chains
    
    ### 4. PyPDF2
    - **Purpose**: PDF text extraction
...
    
...
    
...
    
...
    
...
    
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT token obtained from /api/v1/auth/login endpoint"
  schemas:
...
    ResumeScreeningRequest:
      type: object
      properties:
        job_id:
          type: integer
          example: 5
          description: "Job listing ID to screen candidates for"
        resume_ids:
          type: array
          items:
            type: integer
          example: [101, 102, 103]
          description: "Specific application IDs to screen (optional, screens all if not provided)"
        job_description:
          type: string
          description: "Override job description from listing (optional)"
      required:
        - job_id
    ResumeScreeningResult:
      type: object
      properties:
        candidate_name:
          type: string
          example: "Jane Smith"
        application_id:
          type: integer
          example: 101
        overall_fit_score:
          type: integer
          minimum: 0
          maximum: 100
          example: 92
        skills_match:
          type: array
          items:
            type: string
          example: ["Python (Expert)", "React (Advanced)", "AWS (Intermediate)"]
        experience_match:
          type: string
          example: "8 years - Excellent match"
        education:
          type: string
          example: "B.S. Computer Science, MIT"
        strengths:
          type: array
          items:
            type: string
          example: ["Strong technical leadership", "Cloud architecture expertise"]
        gaps:
          type: array
          items:
            type: string
          example: ["Limited DevOps experience"]
        summary:
          type: string
          example: "Excellent candidate with strong technical background and relevant experience..."
    ResumeScreeningResultResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        job_id:
          type: integer
          example: 5
        job_title:
          type: string
          example: "Senior Software Engineer"
        total_analyzed:
          type: integer
          example: 15
        average_score:
          type: number
          format: float
          example: 78.5
        top_candidate:
          type: string
          example: "Jane Smith"
        analysis_id:
          type: string
          example: "abc123def456"
          description: "Unique ID for retrieving results later"
        results:
          type: array
          items:
            $ref: '#/components/schemas/ResumeScreeningResult'
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/ai/resume-screener/screen:
    post:
      tags:
        - AI - Resume Screener
      summary: Screen Resumes with AI (GenAI)
      description: |
        **User Story #5: Resume Screening**
        Reduce manual effort in screening resumes by automatically analyzing candidates using AI.
        
        **GenAI Integration Details:**
        - **Model:** Google Gemini 1.5 Flash
        - **PDF Processing:** PyPDF2 for text extraction
        - **Analysis:** AI-powered skill matching and experience validation
        
        **Features:**
        - Intelligent resume matching against job description
        - Skill detection with proficiency levels
        - Experience level verification
        - Education qualification validation
        - 0-100 scoring system with detailed breakdown
        - Batch processing support
        - Permanent storage with unique analysis ID
        - Top candidate identification
        
        **How it Works:**
        1. Extract text from PDF resumes using PyPDF2
        2. Analyze each resume against job description using Gemini
        3. Score candidates on multiple dimensions (skills, experience, education)
        4. Generate detailed analysis with strengths and gaps
        5. Store results permanently with unique analysis ID
        
        **Error Handling:**
        - 400: Invalid job ID or no resumes found
        - 403: HR access required
        - 404: Job listing or applications not found
        - 500: AI processing error
        
        **Access:** HR only
        **Performance:** Typically 3-5 seconds per resume
      operationId: screenResumes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeScreeningRequest'
      responses:
        '200':
          description: Resumes screened successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeScreeningResultResponse'
        '400':
          description: Bad Request
        '403':
          description: Forbidden - HR access required
        '404':
          description: Not Found - Job listing or applications not found
  /api/v1/ai/resume-screener/results/{analysis_id}:
    get:
      tags:
        - AI - Resume Screener
      summary: Get Saved Screening Results
      description: |
        Retrieve previously saved resume screening results by analysis ID.
        
        **Access:** HR only
      operationId: getScreeningResults
      parameters:
        - name: analysis_id
          in: path
          required: true
          schema:
            type: string
          description: Unique analysis ID from screening operation
      responses:
        '200':
          description: Results retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeScreeningResultResponse'
        '404':
          description: Analysis not found
  /api/v1/ai/resume-screener/history:
    get:
      tags:
        - AI - Resume Screener
      summary: Get Screening History
      description: |
        Get history of all resume screening analyses.
        Optionally filter by job_id.
        
        **Access:** HR only
      operationId: getScreeningHistory
      parameters:
        - name: job_id
          in: query
          schema:
            type: integer
          description: Filter by job listing ID (optional)
      responses:
        '200':
          description: History retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  total:
                    type: integer
                  history:
                    type: array
                    items:
                      type: object
...